MVVMIFY: AddChild page logic moved to ViewModel
